+++ sat/config.py
@@ -43,6 +43,11 @@ class ConfigValidationError(Exception):
     pass
 
 
+class ConfigFileExistsError(Exception):
+    """The configuration file already exists."""
+    pass
+
+
 def validate_log_level(level):
     """Validates the given log level.
 
@@ -350,7 +355,7 @@ def process_toml_output(toml_str):
     """
     copyright_stmt = """\
         Default configuration file for SAT.
-        (C) Copyright 2019-2020 Hewlett Packard Enterprise Development LP.
+        (C) Copyright 2019-2021 Hewlett Packard Enterprise Development LP.
 
         Permission is hereby granted, free of charge, to any person obtaining a
         copy of this software and associated documentation files (the "Software"),
@@ -377,38 +382,39 @@ def process_toml_output(toml_str):
                                               not (line.startswith('username') and not line.endswith('""\n'))]))
 
 
-def generate_default_config(path=None, username=None, force=False, log_when_file_exists=False):
+def generate_default_config(path, username=None, force=False):
     """Generates a default SAT configuration file.
 
     This function generates a default SAT configuration file if one does not
-    already exist.  It generates the configuration file in the default
-    location or the location specified by $SAT_CONFIG_FILE.
+    already exist.  It generates the configuration file at the path specified
+    by 'path'.
 
     Args:
         path(str): write the configuration file to the given path.
         username(str): generate the configuration file with
             api_gateway.username set to the given username.
         force(bool): if True, overwrite existing configuration files.
-        log_when_file_exists(bool): if True, log that the log file exists
-            at ERROR level, else log at DEBUG level.
 
     Returns:
         None
+
+    Raises:
+        ConfigFileExistsError: if the configuration file already exists.
+        SystemExit: if unable to create the configuration directory or unable
+            to create the configuration file.
     """
-    # Allow the user to specify an alternate config file in an env variable
-    config_file_path = path or os.getenv('SAT_CONFIG_FILE', DEFAULT_CONFIG_PATH)
-    if os.path.isfile(config_file_path) and not force:
-        log_fn = LOGGER.error if log_when_file_exists else LOGGER.debug
-        log_fn(f'Configuration file "{config_file_path}" already exists.  Not generating configuration file.')
-        return
-    config_file_dir = os.path.dirname(config_file_path)
+    if os.path.isfile(path) and not force:
+        raise ConfigFileExistsError(
+            f'Configuration file "{path}" already exists. Not generating configuration file.'
+        )
+    config_file_dir = os.path.dirname(path)
     if not os.path.isdir(config_file_dir):
         try:
             os.makedirs(config_file_dir, exist_ok=True)
         except OSError as e:
             LOGGER.error(f'Unable to create directory {config_file_dir}: {e}')
             raise SystemExit(1)
-    LOGGER.info(f'Creating default configuration file at {config_file_path}')
+    LOGGER.info(f'Creating default configuration file at {path}')
 
     # Convert SAT_CONFIG_SPEC to a dict
     config_spec = {
@@ -423,7 +429,7 @@ def generate_default_config(path=None, username=None, force=False, log_when_file
         config_spec['api_gateway']['username'] = username
 
     try:
-        output_stream = open(config_file_path, 'w')
+        output_stream = open(path, 'w')
         with output_stream:
             toml_str = toml.dumps(config_spec)
             output_stream.write(process_toml_output(toml_str))